feat: reflect the selected method in the url
