+++ src/app/page.tsx
@@ -5,7 +5,7 @@ export default function Home() {
     return (
         <Box sx={{ margin: '0 auto', textAlign: 'center' }}>
             <Typography variant="h1">GraphQL Task</Typography>
-            <Link href="/restful">
+            <Link href="/get">
                 <Button variant="contained">Restful</Button>
             </Link>
         </Box>
